UI v0.8
Mobile responsive update:
- Hide case filter element for mobile user
- Allow mobile user to use case filtering by popup

Bug Appeared:
- The popup auto reset the filter after closed
- Therefore the filter only work for once until mobile user refresh the page

Problem could solve by:
- Make the cases filter component return the user selected filter value
- Get and Set filter value on raspberry detail component
- If selected filter values are not empty then the inputs / dropdowns should default by the selected filter values
