+++ tlc-website copy/src/components/FilterCase/FilterCase.css
@@ -1,0 +1,39 @@
+.filter-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+    background-color: #4a4e69;
+    padding: 10px;
+    border-radius: 5px;
+    margin-bottom: 20px;
+  }
+  
+  /* The screen is smallest than 700px*/
+@media only screen and (max-width: 700px) {
+    /* Hide filter  */
+    .filter-container {
+      display: none;
+    }
+}
+  
+.filter-item {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    margin: 0 10px;
+  
+    .label {
+      font-size: 14px;
+      color: #edf2f4;
+      margin-bottom: 5px;
+    }
+  
+    .input,
+    .select {
+      padding: 5px;
+      border-radius: 4px;
+      border: 1px solid #ccc;
+      font-size: 14px;
+    }
+}
