+++ tlc-website copy/src/components/RaspiDetail/RaspiDetail.js
@@ -4,11 +4,16 @@ import { getStorage, ref, listAll, deleteObject } from "firebase/storage";
 import { db } from "../../firebase/firebase.js"; // Update your Firebase paths
 import { getAuth } from "firebase/auth";
 import "./RaspiDetail.css";
-import { GrLinkPrevious, FaImage, FaFilm, MdOutlineWork, MdDelete } from "../../images/Icons.js"
 import CasePreview from "../CasePreview/CasePreview.js"
 import CaseAdd from "../CaseAdd/CaseAdd"
 import Film from "../Film/Film.js"
 import Album from "../Album/Album.js"
+import Filter from "../FilterCase/FilterCase.js"
+import { 
+            GrLinkPrevious, FaImage, FaFilm, 
+            MdOutlineWork, MdDelete, MdFilterListAlt 
+        } from "../../images/Icons.js"
+
 
 const RaspiDetail = ({pi, onBack}) => {
     const [timeLapseCases, setTimeLapseCases] = useState([]);
@@ -22,13 +27,7 @@ const RaspiDetail = ({pi, onBack}) => {
     const [isModalOpen, setIsModalOpen] = useState(false);
     const [caseToDelete, setCaseToDelete] = useState(null);
     const [filteredCases, setFilteredCases] = useState([]);
-    const [filters, setFilters] = useState({
-                                            name: "",
-                                            status: "",
-                                            resolution: "",
-                                            startDate: "",
-                                            endDate: "",
-                                        });
+    const [openFilter, setOpenFilter] = useState(false);
     const auth = getAuth();
     const currentUser = auth.currentUser;
     const userUID = currentUser ? currentUser.uid : null;
@@ -67,27 +66,6 @@ const RaspiDetail = ({pi, onBack}) => {
         const seconds = date.getSeconds().toString().padStart(2, "0");
         return `${day}-${month}-${year} at ${hours}:${minutes}:${seconds}`;
     };
-      
-    // const handleSaveSuccess = async (id = null) => {
-    //     const timeLapseRef = collection(db, `raspberrys/${pi.serial}/TimeLapseCase`);
-    //     const timeLapseSnapshot = await getDocs(timeLapseRef);
-    
-    //     const updatedCases = timeLapseSnapshot.docs.map((doc) => ({
-    //         id: doc.id,
-    //         ...doc.data(),
-    //     }));
-    
-    //     setTimeLapseCases(updatedCases); // Update the list of cases
-
-    //     if (selectedCaseId) {
-    //         const updatedCase = updatedCases.find((caseItem) => caseItem.id === selectedCaseId.id);
-    //         setSelectedCaseId(updatedCase); // Update the selected case
-    //         setShowPreviewPage(true); // Stay in edit mode
-    //     } else {
-    //         setShowPreviewPage(false);
-    //         setShowCaseAddPage(false);
-    //     }
-    // };
     
     if (showPreviewPage) {
         return (
@@ -242,37 +220,32 @@ const RaspiDetail = ({pi, onBack}) => {
         );
         };
 
-    // Handle Filtering
-    const handleFilterChange = (filterType, value) => {
-        const updatedFilters = { ...filters, [filterType]: value };
-        setFilters(updatedFilters);
     
-        const filtered = timeLapseCases.filter((caseItem) => {
-            const matchesName = updatedFilters.name
-                ? caseItem.name.toLowerCase().includes(updatedFilters.name.toLowerCase())
-                : true;
-        
-            const matchesStatus = updatedFilters.status
-                ? caseItem.status === updatedFilters.status
-                : true;
-        
-            const matchesResolution = updatedFilters.resolution
-                ? caseItem.resolution === updatedFilters.resolution
-                : true;
+const FilterPopUp = ({ open, onClose }) => {
+    if(!open) return null;
+    // console.log(open);
 
-            const caseStartDate = new Date(caseItem.caseStart);
-            const caseEndDate = new Date(caseItem.caseEnd);
+    return (
+        <div className="modal-overlay">
+            <div className="modal-content">
+                <Filter 
+                    cases = {timeLapseCases} 
+                    fCases = {(filtered) => {
+                        // console.log(filtered);
+                        setFilteredCases(filtered);
+                    }} 
+                />
+                <button 
+                    className="close-filter"
+                    onClick={onClose}
+                >
+                        Close
+                </button>
+            </div>
+        </div>
+    );
+};
 
-            const matchesDateRange =
-            (!updatedFilters.startDate || caseStartDate >= new Date(updatedFilters.startDate)) &&
-            (!updatedFilters.endDate || caseEndDate <= new Date(updatedFilters.endDate));      
-    
-          return matchesName && matchesStatus && matchesResolution && matchesDateRange;
-        });
-    
-        setFilteredCases(filtered);
-      };
-    
 return (
     <header>
         {/* <p><strong>Serial Number:</strong> {pi.serial}</p> */}
@@ -296,66 +269,25 @@ return (
 
                 <p className="device-name"><strong>{pi.data.NAME}</strong></p>
                 {/* Filter Input Field */}
-                <div className="filter-container">
-                    <div class="filter-item">
-                        <label for="filter-name">Name:</label>
-                        <input
-                            type="text"
-                            placeholder="Filter by name..."
-                            value={filters.name}
-                            onChange={(e) => handleFilterChange("name", e.target.value)}
-                        />
-                    </div>
+                <div className="filter-section">
+                    <Filter 
+                        cases = {timeLapseCases} 
+                        fCases = {(filtered) => {
+                            // console.log(filtered);
+                            setFilteredCases(filtered);
+                        }} 
+                    />
 
-                    <div class="filter-item">
-                        <label for="filter-status">Status:</label>
-                        <select
-                            value={filters.status || ""}
-                            onChange={(e) => handleFilterChange("status", e.target.value)}
-                            >
-                            <option value="">All</option>
-                            <option value="running">Running</option>
-                            <option value="standby">Standby</option>
-                            <option value="completed">Completed</option>
-                            <option value="aborted">Aborted</option>
-                        </select>
+                    <div className="small-screen-filter">
+                        <button onClick={() => setOpenFilter(true)}>
+                            <MdFilterListAlt />
+                        </button>
                     </div>
-                    
-                    <div class="filter-item">
-                        <label for="filter-status">Resolution:</label>
-                            <select
-                            value={filters.resolution || ""}
-                            onChange={(e) => handleFilterChange("resolution", e.target.value)}
-                            >
-                            <option value="">All</option>
-                            <option value="Max_View">12MP (4056x3040)</option>
-                            <option value="4K_UHD">4K UHD (3840x2160)</option>
-                            <option value="2K_UHD">2K UHD (2560x1440)</option>
-                            <option value="1080p">1080p Full HD (1920x1080)</option>
-                            <option value="720p">720p HD (1280x720)</option>
-                            <option value="SD_480p">480p SD (640x480)</option>
-                        </select>
-                    </div>
-
-                    <div class="filter-item">
-                        <label for="filter-status">Start:</label>
-                            <input
-                            type="date"
-                            value={filters.startDate || ""}
-                            onChange={(e) => handleFilterChange("startDate", e.target.value)}
-                        />
-                    </div>
-
-                    <div class="filter-item">
-                        <label for="filter-status">End:</label>
-                            <input
-                            type="date"
-                            value={filters.endDate || ""}
-                            onChange={(e) => handleFilterChange("endDate", e.target.value)}
-                        />
-                    </div>
-
-
+            
+                    <FilterPopUp 
+                        open={openFilter}
+                        onClose={() => setOpenFilter(false)}
+                    />
                 </div>
 
                 {filteredCases.map((timeLapseCase) => (
